+++ pyne/stlcontainers.pyx
@@ -809,14 +809,14 @@ cdef class _MapStrUInt:
             for key, value in new_map.items():
                 key_bytes = key.encode()
 
-                item = pair[std_string, extra_types.uint32](std_string(<char *> key_bytes), <extra_types.uint32> long(value))
+                item = pair[std_string, extra_types.uint32](std_string(<char *> key_bytes), <extra_types.uint32> value)
                 self.map_ptr.insert(item)
         elif hasattr(new_map, '__len__'):
             self.map_ptr = new cpp_map[std_string, extra_types.uint32]()
             for key, value in new_map:
                 key_bytes = key.encode()
 
-                item = pair[std_string, extra_types.uint32](std_string(<char *> key_bytes), <extra_types.uint32> long(value))
+                item = pair[std_string, extra_types.uint32](std_string(<char *> key_bytes), <extra_types.uint32> value)
                 self.map_ptr.insert(item)
         elif bool(new_map):
             self.map_ptr = new cpp_map[std_string, extra_types.uint32]()
@@ -872,7 +872,7 @@ cdef class _MapStrUInt:
         cdef pair[std_string, extra_types.uint32] item
         key_bytes = key.encode()
 
-        item = pair[std_string, extra_types.uint32](std_string(<char *> key_bytes), <extra_types.uint32> long(value))
+        item = pair[std_string, extra_types.uint32](std_string(<char *> key_bytes), <extra_types.uint32> value)
         if 0 < self.map_ptr.count(std_string(<char *> key_bytes)):
             self.map_ptr.erase(std_string(<char *> key_bytes))
         self.map_ptr.insert(item)
@@ -960,14 +960,14 @@ cdef class _MapUIntStr:
             for key, value in new_map.items():
 
                 value_bytes = value.encode()
-                item = pair[extra_types.uint32, std_string](<extra_types.uint32> long(key), std_string(<char *> value_bytes))
+                item = pair[extra_types.uint32, std_string](<extra_types.uint32> key, std_string(<char *> value_bytes))
                 self.map_ptr.insert(item)
         elif hasattr(new_map, '__len__'):
             self.map_ptr = new cpp_map[extra_types.uint32, std_string]()
             for key, value in new_map:
 
                 value_bytes = value.encode()
-                item = pair[extra_types.uint32, std_string](<extra_types.uint32> long(key), std_string(<char *> value_bytes))
+                item = pair[extra_types.uint32, std_string](<extra_types.uint32> key, std_string(<char *> value_bytes))
                 self.map_ptr.insert(item)
         elif bool(new_map):
             self.map_ptr = new cpp_map[extra_types.uint32, std_string]()
@@ -985,7 +985,7 @@ cdef class _MapUIntStr:
         if not isinstance(key, int):
             return False
 
-        k = <extra_types.uint32> long(key)
+        k = <extra_types.uint32> key
 
         if 0 < self.map_ptr.count(k):
             return True
@@ -1008,7 +1008,7 @@ cdef class _MapUIntStr:
         if not isinstance(key, int):
             raise TypeError("Only unsigned integer keys are valid.")
 
-        k = <extra_types.uint32> long(key)
+        k = <extra_types.uint32> key
 
         if 0 < self.map_ptr.count(k):
             v = deref(self.map_ptr)[k]
@@ -1023,9 +1023,9 @@ cdef class _MapUIntStr:
         cdef pair[extra_types.uint32, std_string] item
 
         value_bytes = value.encode()
-        item = pair[extra_types.uint32, std_string](<extra_types.uint32> long(key), std_string(<char *> value_bytes))
-        if 0 < self.map_ptr.count(<extra_types.uint32> long(key)):
-            self.map_ptr.erase(<extra_types.uint32> long(key))
+        item = pair[extra_types.uint32, std_string](<extra_types.uint32> key, std_string(<char *> value_bytes))
+        if 0 < self.map_ptr.count(<extra_types.uint32> key):
+            self.map_ptr.erase(<extra_types.uint32> key)
         self.map_ptr.insert(item)
 
     def __delitem__(self, key):
@@ -1033,7 +1033,7 @@ cdef class _MapUIntStr:
 
         if key in self:
 
-            k = <extra_types.uint32> long(key)
+            k = <extra_types.uint32> key
             self.map_ptr.erase(k)
 
 
@@ -1262,14 +1262,14 @@ cdef class _MapUIntUInt:
             for key, value in new_map.items():
 
 
-                item = pair[extra_types.uint32, extra_types.uint32](<extra_types.uint32> long(key), <extra_types.uint32> long(value))
+                item = pair[extra_types.uint32, extra_types.uint32](<extra_types.uint32> key, <extra_types.uint32> value)
                 self.map_ptr.insert(item)
         elif hasattr(new_map, '__len__'):
             self.map_ptr = new cpp_map[extra_types.uint32, extra_types.uint32]()
             for key, value in new_map:
 
 
-                item = pair[extra_types.uint32, extra_types.uint32](<extra_types.uint32> long(key), <extra_types.uint32> long(value))
+                item = pair[extra_types.uint32, extra_types.uint32](<extra_types.uint32> key, <extra_types.uint32> value)
                 self.map_ptr.insert(item)
         elif bool(new_map):
             self.map_ptr = new cpp_map[extra_types.uint32, extra_types.uint32]()
@@ -1287,7 +1287,7 @@ cdef class _MapUIntUInt:
         if not isinstance(key, int):
             return False
 
-        k = <extra_types.uint32> long(key)
+        k = <extra_types.uint32> key
 
         if 0 < self.map_ptr.count(k):
             return True
@@ -1310,7 +1310,7 @@ cdef class _MapUIntUInt:
         if not isinstance(key, int):
             raise TypeError("Only unsigned integer keys are valid.")
 
-        k = <extra_types.uint32> long(key)
+        k = <extra_types.uint32> key
 
         if 0 < self.map_ptr.count(k):
             v = deref(self.map_ptr)[k]
@@ -1325,9 +1325,9 @@ cdef class _MapUIntUInt:
         cdef pair[extra_types.uint32, extra_types.uint32] item
 
 
-        item = pair[extra_types.uint32, extra_types.uint32](<extra_types.uint32> long(key), <extra_types.uint32> long(value))
-        if 0 < self.map_ptr.count(<extra_types.uint32> long(key)):
-            self.map_ptr.erase(<extra_types.uint32> long(key))
+        item = pair[extra_types.uint32, extra_types.uint32](<extra_types.uint32> key, <extra_types.uint32> value)
+        if 0 < self.map_ptr.count(<extra_types.uint32> key):
+            self.map_ptr.erase(<extra_types.uint32> key)
         self.map_ptr.insert(item)
 
     def __delitem__(self, key):
@@ -1335,7 +1335,7 @@ cdef class _MapUIntUInt:
 
         if key in self:
 
-            k = <extra_types.uint32> long(key)
+            k = <extra_types.uint32> key
             self.map_ptr.erase(k)
 
 
@@ -1866,14 +1866,14 @@ cdef class _MapUIntDouble:
             for key, value in new_map.items():
 
 
-                item = pair[extra_types.uint32, double](<extra_types.uint32> long(key), <double> value)
+                item = pair[extra_types.uint32, double](<extra_types.uint32> key, <double> value)
                 self.map_ptr.insert(item)
         elif hasattr(new_map, '__len__'):
             self.map_ptr = new cpp_map[extra_types.uint32, double]()
             for key, value in new_map:
 
 
-                item = pair[extra_types.uint32, double](<extra_types.uint32> long(key), <double> value)
+                item = pair[extra_types.uint32, double](<extra_types.uint32> key, <double> value)
                 self.map_ptr.insert(item)
         elif bool(new_map):
             self.map_ptr = new cpp_map[extra_types.uint32, double]()
@@ -1891,7 +1891,7 @@ cdef class _MapUIntDouble:
         if not isinstance(key, int):
             return False
 
-        k = <extra_types.uint32> long(key)
+        k = <extra_types.uint32> key
 
         if 0 < self.map_ptr.count(k):
             return True
@@ -1914,7 +1914,7 @@ cdef class _MapUIntDouble:
         if not isinstance(key, int):
             raise TypeError("Only unsigned integer keys are valid.")
 
-        k = <extra_types.uint32> long(key)
+        k = <extra_types.uint32> key
 
         if 0 < self.map_ptr.count(k):
             v = deref(self.map_ptr)[k]
@@ -1929,9 +1929,9 @@ cdef class _MapUIntDouble:
         cdef pair[extra_types.uint32, double] item
 
 
-        item = pair[extra_types.uint32, double](<extra_types.uint32> long(key), <double> value)
-        if 0 < self.map_ptr.count(<extra_types.uint32> long(key)):
-            self.map_ptr.erase(<extra_types.uint32> long(key))
+        item = pair[extra_types.uint32, double](<extra_types.uint32> key, <double> value)
+        if 0 < self.map_ptr.count(<extra_types.uint32> key):
+            self.map_ptr.erase(<extra_types.uint32> key)
         self.map_ptr.insert(item)
 
     def __delitem__(self, key):
@@ -1939,7 +1939,7 @@ cdef class _MapUIntDouble:
 
         if key in self:
 
-            k = <extra_types.uint32> long(key)
+            k = <extra_types.uint32> key
             self.map_ptr.erase(k)